backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.4.0
PyPDF2==3.0.1
pypdf==4.0.1
pdf2image==1.17.0
Pillow==10.4.0
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27
boto3==1.34.34
httpx==0.26.0
tenacity==8.2.3
structlog==24.4.0
sentry-sdk==2.0.0
prometheus-client==0.21.0
pydantic-extra-types==2.9.0
email-validator==2.2.0
bcrypt==4.1.2
